refactor(plan): extract change vocabulary
