+++ tests/domain/plan/test_diff.py
@@ -32,7 +32,7 @@ from delta_engine.domain.plan.actions import (
     UnsetProperty,
     UnsetTableTag,
 )
-from delta_engine.domain.plan.diff import (
+from delta_engine.domain.plan.changes import (
     ColumnAdded,
     ColumnCommentChanged,
     ColumnDataTypeChanged,
@@ -50,10 +50,12 @@ from delta_engine.domain.plan.diff import (
     PropertyUndeclared,
     PropertyUnset,
     TableCommentChanged,
-    TableDrift,
-    TableMissing,
     TableTagSet,
     TableTagUnset,
+)
+from delta_engine.domain.plan.diff import (
+    TableDrift,
+    TableMissing,
     diff_table,
 )
